convert observations to float16

+++ craftax/craftax_marl_basic/envs/craftax_symbolic_env.py
@@ -78,7 +78,7 @@ class CraftaxMARLSymbolicEnv(MultiAgentEnv):
                 state, 
                 self.static_env_params,
             )
-        )
+        ).astype(jnp.float16)
         obs = {n:o for n,o in zip(self.agents, obs_sym)}
         return obs
